+++ core/src/test/java/com/graphhopper/reader/osm/OSMReaderTest.java
@@ -398,6 +398,36 @@ public class OSMReaderTest {
     }
 
     @Test
+    public void testFords() {
+        CarFlagEncoder car = new CarFlagEncoder(new PMap("block_fords=true"));
+        GraphHopper hopper = new GraphHopper();
+        hopper.getEncodingManagerBuilder().add(car);
+        hopper.setOSMFile(getClass().getResource("test-barriers3.xml").getFile()).
+                setGraphHopperLocation(dir).
+                setProfiles(
+                        new Profile("car").setVehicle("car").setWeighting("fastest")
+                ).
+                setMinNetworkSize(0).
+                importOrLoad();
+        Graph graph = hopper.getGraphHopperStorage();
+        // our way is split into five edges, because there are two ford nodes
+        assertEquals(5, graph.getEdges());
+        FlagEncoder encoder = hopper.getEncodingManager().fetchEdgeEncoders().get(0);
+        int blocked = 0;
+        int notBlocked = 0;
+        AllEdgesIterator edge = graph.getAllEdges();
+        while (edge.next()) {
+            if (!edge.get(encoder.getAccessEnc()))
+                blocked++;
+            else
+                notBlocked++;
+        }
+        // two blocked edges and three accessible edges
+        assertEquals(2, blocked);
+        assertEquals(3, notBlocked);
+    }
+
+    @Test
     public void avoidsLoopEdges_1525() {
         // loops in OSM should be avoided by adding additional tower node (see #1525, #1531)
         //     C - D